updated to version 1.2
NOTE: This sketch is only required for the older Adafruit 1651 TFT touchscreen displays that incorporate the now obsolete STMPE610 touchscreen controller.  The newer Adafruit 1651 TFT touchscreen displays that now incorporate the newer TSP2007 touchscreen controller should not need this sketch (and it is not written to support the newer touchscreen controller).